Move to TypeScript (#45)
