+++ .github/workflows/publish.yml
@@ -13,6 +13,7 @@ jobs:
         with:
           node-version: "16.x"
           registry-url: "https://registry.npmjs.org"
+      - run: npx tsc
       - run: npm ci
       - run: npm publish --access=public
         env:
