+++ .gitignore
@@ -43,6 +43,7 @@ jspm_packages/
 
 # TypeScript v1 declaration files
 typings/
+lib/
 
 # TypeScript cache
 *.tsbuildinfo
