+++ .prettierrc.json
@@ -1,0 +1,7 @@
+{
+    "trailingComma": "none",
+    "tabWidth": 4,
+    "semi": true,
+    "useTabs": true,
+    "bracketSpacing": true
+}
