+++ index.ts
@@ -1,0 +1,17 @@
+/** Modules **/
+
+export { flip } from "./src/flip";
+export { log } from "./src/log";
+export { print } from "./src/print";
+export { random } from "./src/random";
+export { uuid } from "./src/uuid";
+
+/** Util **/
+
+export { boolToNum } from "./src/types/boolToNum";
+export { boolToStr } from "./src/types/boolToStr";
+export { numToBool } from "./src/types/numToBool";
+export { numToStr } from "./src/types/numToStr";
+export { removeDuplicates } from "./src/util/removeDuplicates";
+export { strToBool } from "./src/types/strToBool";
+export { strToNum } from "./src/types/strToNum";
