--- src/index.js
@@ -1,3 +1,0 @@
-const modules = require("./modules");
-
-module.exports = modules;
