--- src/modules/util/boolToStr.js
@@ -1,13 +1,0 @@
-/**
- * @function `boolToStr` - Convert a boolean to a string.
- * @param { Boolean } boolean - *Required* The boolean to be converted.
- * @returns { String } - The output of the conversion.
- */
-
-function boolToStr(boolean) {
-    if (typeof boolean !== "boolean") throw new Error("No boolean provided");
-
-    return boolean.toString();
-}
-
-module.exports = boolToStr;
