--- src/modules/util/numToStr.js
@@ -1,13 +1,0 @@
-/**
- * @function `numToStr` - Convert a number to a string.
- * @param { Number } number - *Required* The number to be converted.
- * @returns { String } - The output of the conversion.
- */
-
-function numToStr(number) {
-    if (typeof number !== "number") throw new Error("No number provided");
-
-    return number.toString();
-}
-
-module.exports = numToStr;
