+++ src/types/boolToNum.ts
@@ -1,0 +1,5 @@
+export function boolToNum(boolean: boolean): number {
+  if (typeof boolean !== "boolean") throw new Error("No boolean provided");
+
+  return boolean ? 1 : 0;
+}
