+++ src/types/boolToStr.ts
@@ -1,0 +1,5 @@
+export function boolToStr(boolean: boolean): string {
+	if (typeof boolean !== "boolean") throw new Error("No boolean provided");
+
+	return boolean.toString();
+}
