+++ src/types/strToBool.ts
@@ -1,0 +1,8 @@
+export function strToBool(string: string): boolean {
+	if (!string || typeof string !== "string")
+		throw new Error("No string provided");
+	if (string !== "true" && string !== "false")
+		throw new Error("No valid string provided");
+
+	return string === "true";
+}
