+++ src/types/strToNum.ts
@@ -1,0 +1,7 @@
+export function strToNum(string: string): number {
+	if (!string || typeof string !== "string")
+		throw new Error("No string provided");
+	if (isNaN(Number(string))) throw new Error("No valid string provided");
+
+	return Number(string);
+}
